% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ACS_CP.R
\docType{data}
\name{ACS_CP}
\alias{ACS_CP}
\title{ACS_CP}
\format{
A data frame with 1842 rows and 4 variables:
\describe{
  \item{age}{age, in years}
  \item{ind}{industry code, in four digits}
  \item{baplus}{1 if a respondent has a master’s degree, a professional degree, or a doctoral degree; 0 otherwise}
  \item{topincome}{1 if an observation belongs to the case sample; NA otherwise}
}
}
\source{
\url{https://usa.ipums.org/usa/}
}
\usage{
ACS_CP
}
\description{
A case-population sample extracted from American Community Survey (ACS) 2018, restricted to white males residing in California with at least a bachelor's degree.
The original ACS dataset is not from case-population sampling, but this case-population sample is obtained by the following procedure.
The case sample is composed of 921 individuals whose income is top-coded.
The control sample of equal size is randomly drawn without replacement from all observations and its top-coded status is coded missing.
Age is restricted to be between 25 and 70.
}
\keyword{datasets}
