% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicc.R
\name{cicc}
\alias{cicc}
\title{Upper bound on the average of log relative risk}
\usage{
cicc(y, t, x, p)
}
\arguments{
\item{y}{n-dimensional vector of binary outcomes}

\item{t}{n-dimensional vector of binary treatments}

\item{x}{n by p matrix of matrix of covariates}

\item{p}{the upper bound for the unknown true case probability}
}
\value{
estimate and standard error of the upper bound on the average of log relative risk
}
\description{
Upper bound on the average of log relative risk
}
