% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicc_AR.R
\name{cicc_AR}
\alias{cicc_AR}
\title{Causal Inference for Attributable Risk}
\usage{
cicc_AR(
  y,
  t,
  x,
  sampling = "cc",
  p_upper = 1L,
  cov_prob = 0.95,
  length = 21L,
  interaction = FALSE,
  no_boot = 999L
)
}
\arguments{
\item{y}{n-dimensional vector of binary outcomes}

\item{t}{n-dimensional vector of binary treatments}

\item{x}{n by p matrix of covariates}

\item{sampling}{'cc' for case-control sampling; 'cp' for case-population sampling (default sampling =  'cc')}

\item{p_upper}{a specified upper bound for the unknown true case probability (default = 1)}

\item{cov_prob}{parameter for coverage probability of a confidence interval (default = 0.95)}

\item{length}{specified length of a sequence from 0 to p_upper (default = 21)}

\item{interaction}{TRUE if there are interaction terms in the retrospective logistic model; FALSE if not (default = FALSE)}

\item{no_boot}{number of bootstrap repetitions to compute the confidence intervals (default = 999)}
}
\value{
An S3 object of type "ciccr". The object has the following elements:
\item{est}{(length)-dimensional vector of the upper bounds on the average of attributable risk}
\item{se}{(length)-dimensional vector of pointwise standard errors}
\item{ci}{(length)-dimensional vector of the upper ends of pointwise confidence intervals}
\item{pseq}{(length)-dimensional vector of a grid from 0 to p_upper}
}
\description{
Provides an upper bound on the average of attributable risk.
}
\examples{
# use the ACS dataset included in the package.
  y = ciccr::ACS$topincome
  t = ciccr::ACS$baplus
  x = ciccr::ACS$age
  AR = cicc_AR(y, t, x, sampling = 'cc', no_boot = 100)

}
\references{
Sung Jae Jun and Sokbae Lee. Causal Inference in Case-Control Studies.
\url{https://arxiv.org/abs/2004.08318}.
}
