% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicc_plot.R
\name{cicc_plot}
\alias{cicc_plot}
\title{Upper Bounds on Relative Risk and Attributable Risk}
\usage{
cicc_plot(
  y,
  t,
  x,
  sampling = "cc",
  p_upper = 1L,
  cov_prob = 0.95,
  length = 20L,
  interaction = FALSE,
  no_boot = 0L,
  save_plots = FALSE,
  file_name = Sys.Date(),
  plots_ctl = 0.3,
  plots_dir = FALSE
)
}
\arguments{
\item{y}{n-dimensional vector of binary outcomes}

\item{t}{n-dimensional vector of binary treatments}

\item{x}{n by p matrix of covariates}

\item{sampling}{'cc' for case-control sampling; 'cp' for case-population sampling (default sampling =  'cc')}

\item{p_upper}{a specified upper bound for the unknown true case probability (default = 1; relevant for only 'cc')}

\item{cov_prob}{parameter for coverage probability of a confidence interval (default = 0.95)}

\item{length}{specified length of a sequence from 0 to p_upper (default = 20; ; relevant for only 'cc')}

\item{interaction}{TRUE if there are interaction terms in the retrospective logistic model
to estimate the upper bound on attributable risk; FALSE if not (default = FALSE)}

\item{no_boot}{number of bootstrap repetitions to compute the confidence intervals
for attributable risk (default = 0)}

\item{save_plots}{TRUE if the plots are saved as pdf files; FALSE if not (default = FALSE)}

\item{file_name}{the pdf file name to save the plots (default = Sys.Date())}

\item{plots_ctl}{value to determine the topleft position of the legend in the figure
a large value makes the legend far away from the confidence intervals (default = 0.3)}

\item{plots_dir}{a directory where the plots are saved (default = FALSE);
plots will be saved under "(current working directory)/figures" by default.}
}
\value{
A X-Y plot where the X axis shows the range of p from 0 to p_upper and
the Y axis depicts both point estimates and the upper end point of the one-sided confidence intervals.
}
\description{
Plots upper bounds on relative risk and attributable risk
}
\examples{
# use the ACS dataset included in the package.
  y = ciccr::ACS$topincome
  t = ciccr::ACS$baplus
  x = ciccr::ACS$age
  cicc_plot(y, t, x)

}
\references{
Sung Jae Jun and Sokbae Lee. Causal Inference in Case-Control Studies.
\url{https://arxiv.org/abs/2004.08318}.
}
