% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicc_RR.R
\name{cicc_RR}
\alias{cicc_RR}
\title{Causal Inference on Relative Risk}
\usage{
cicc_RR(
  y,
  t,
  x,
  sampling = "cc",
  p_upper = 1L,
  cov_prob = 0.95,
  length = 21L,
  interaction = FALSE,
  no_boot = 0L
)
}
\arguments{
\item{y}{n-dimensional vector of binary outcomes}

\item{t}{n-dimensional vector of binary treatments}

\item{x}{n by p matrix of covariates}

\item{sampling}{'cc' for case-control sampling; 'cp' for case-population sampling (default sampling =  'cc')}

\item{p_upper}{a specified upper bound for the unknown true case probability (default = 1)}

\item{cov_prob}{parameter for coverage probability of a confidence interval (default = 0.95)}

\item{length}{specified length of a sequence from 0 to p_upper (default = 21)}

\item{interaction}{TRUE if there are interaction terms in the retrospective logistic model; FALSE if not (default = FALSE)}

\item{no_boot}{number of bootstrap repetitions to compute the confidence intervals (default = 0)}
}
\value{
An S3 object of type "ciccr". The object has the following elements:
\item{est}{(length)-dimensional vector of the upper bounds on the average of log relative risk}
\item{ci}{(length)-dimensional vector of the upper ends of pointwise one-sided confidence intervals}
\item{pseq}{(length)-dimensional vector of a grid from 0 to p_upper}
\item{cov_prob}{the nominal coverage probability}
\item{return_code}{status of existence of missing values in bootstrap replications}
Notes. The outputs would be the same across different p values for for case-population sampling.
}
\description{
Provides an upper bound on the average of log relative risk.
}
\examples{
# use the ACS dataset included in the package.
  y = ciccr::ACS$topincome
  t = ciccr::ACS$baplus
  x = ciccr::ACS$age
# The upper bound for Pr(y=1) is set as p_upper = 1 by default
# with the defaulty coverage probability of 0.95 for the default choice of case-control sampling.
  cicc_RR(y, t, x)
# The upper bound for Pr(y=1) is set as p_upper = 0.2
# with the coverage probablity of 0.99 for the case-control sample.
  cicc_RR(y, t, x, sampling = 'cc', p_upper = 0.2, cov_prob = 0.99)
}
\references{
Sung Jae Jun and Sokbae Lee. Causal Inference in Case-Control Studies.
\url{https://arxiv.org/abs/2004.08318}.
}
