% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciccRR.R
\name{ciccRR}
\alias{ciccRR}
\title{Causal Inference on Relative Risk}
\usage{
ciccRR(y, t, x, sampling = "cc", cov_prob = 0.95)
}
\arguments{
\item{y}{n-dimensional vector of binary outcomes}

\item{t}{n-dimensional vector of binary treatments}

\item{x}{n by d matrix of covariates}

\item{sampling}{'cc' for case-control sampling; 'cp' for case-population sampling (default sampling =  'cc')}

\item{cov_prob}{parameter for coverage probability of a uniform confidence band (default = 0.95)}
}
\value{
An S3 object of type "ciccr". The object has the following elements:
\item{est}{estimates of the upper bounds on the average of log relative risk at p=0 and p=1}
\item{se}{pointwise standard errors at p=0 and p=1}
\item{ci}{the upper end points of the uniform confidence band at p=0 and p=1}
\item{pseq}{two end points: p=0 and p=1}
}
\description{
Provides upper bounds on the average of log relative risk
under the monotone treatment response (MTR) and monotone treatment selection (MTS) assumptions.
}
\examples{
# use the ACS dataset included in the package.
  y = ciccr::ACS$topincome
  t = ciccr::ACS$baplus
  x = ciccr::ACS$age

  ciccRR(y, t, x, sampling = 'cc', cov_prob = 0.95)

}
\references{
Sung Jae Jun and Sokbae Lee. Causal Inference in Case-Control Studies.
\url{https://arxiv.org/abs/2004.08318}.
}
