% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_DZ_CC.R
\docType{data}
\name{DZ_CC}
\alias{DZ_CC}
\title{DZ_CC}
\format{
A data frame with 689 rows and 5 variables:
\describe{
  \item{parent_ba}{indicator: at least one college-educated parent}
  \item{private_school}{indicator: attended private school before unversity}
  \item{parent_inc}{parents' monthly income (in 1000s Rs)}
  \item{case_sample}{indicator: attended "Very Selective University" (VSU)}
  \item{control_sample}{indicator: attended the other simply "Selective University" (SU)}
}
}
\source{
\url{https://www.journals.uchicago.edu/doi/suppl/10.1086/701808/suppl_file/2014399data.zip}
}
\usage{
DZ_CC
}
\description{
Case-control sample extracted from Delavande and Zafar (2019).
The sample is composed of 689 students who attended either "Very Selective University" (VSU) or "Selective University" (SU).
}
\references{
Delavande, A. and Zafar, B. (2019). University Choice: The Role of Expected Earnings, Non-Pecuniary Outcomes, and Financial Constraints.
Journal of Political Economy 127(5), 2343-2393.
}
\keyword{datasets}
