% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_pr.R
\name{trim_pr}
\alias{trim_pr}
\title{Trimming the estimates to be strictly between 0 and 1}
\usage{
trim_pr(ps, eps = 1e-08)
}
\arguments{
\item{ps}{n-dimensional vector of estimated probabilities}

\item{eps}{a small constant that determines the trimming of the estimated probabilities.
Specifically, the estimate probability is trimmed to be between eps and 1-eps (default = 1e-8).}
}
\value{
\item{ps_tr}{n-dimensional trimmed estimates}
}
\description{
Trimming the estimates to be strictly between 0 and 1
}
