% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACS}
\alias{ACS}
\title{ACS}
\format{
A data frame with 1842 rows and 4 variables:
\describe{
  \item{age}{age, in years}
  \item{ind}{industry code, in four digits}
  \item{baplus}{1 if BA or higher; 0 otherwise}
  \item{topincome}{1 if income is top-coded; 0 otherwise}
  ...
}
}
\source{
\url{https://usa.ipums.org/usa/}
}
\usage{
ACS
}
\description{
A case-control sample  using ACS
}
\keyword{datasets}
