% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_retro_logit.R
\name{avg_retro_logit}
\alias{avg_retro_logit}
\title{Average of the log odds ratio using retrospective logistic regression}
\usage{
avg_retro_logit(y, t, x, w)
}
\arguments{
\item{y}{n-dimensional vector of binary outcomes}

\item{t}{n-dimensional vector of binary treatments}

\item{x}{n by p matrix of matrix of covariates}

\item{w}{'case' if the average is conditional on the case sample; 'control' if it is conditional on the control sample}
}
\value{
estimate and standard error of the weighted average of the log odds ratio using retrospective logistic regression
}
\description{
Average of the log odds ratio using retrospective logistic regression
}
