% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_RR_logit.R
\name{avg_RR_logit}
\alias{avg_RR_logit}
\title{An Average of the Upper Bound of Causal Log Relative Risk}
\usage{
avg_RR_logit(
  y,
  t,
  x,
  sampling = "cc",
  p_upper = 1L,
  length = 20L,
  interaction = FALSE
)
}
\arguments{
\item{y}{n-dimensional vector of binary outcomes}

\item{t}{n-dimensional vector of binary treatments}

\item{x}{n by p matrix of covariates}

\item{sampling}{'cc' for case-control sampling; 'cp' for case-population sampling (default sampling =  'cc')}

\item{p_upper}{specified upper bound for the unknown true case probability (default = 1)}

\item{length}{specified length of a sequence from 0 to p_upper (default = 20)}

\item{interaction}{TRUE if there are interaction terms in the retrospective logistic model; FALSE if not (default = FALSE)}
}
\value{
An S3 object of type "ciccr". The object has the following elements.
\item{est}{(length)-dimensional vector of the average of the upper bound of causal attributable risk}
\item{pseq}{(length)-dimensional vector of a grid from 0 to p_upper}
}
\description{
Averages the upper bound of causal log relative risk using logistic regression models
under the monotone treatment response (MTR) and monotone treatment selection (MTS) assumptions.
}
\examples{
# use the ACS dataset included in the package
  y = ciccr::ACS$topincome
  t = ciccr::ACS$baplus
  x = ciccr::ACS$age
  results = avg_RR_logit(y, t, x, sampling = 'cc')

}
\references{
Sung Jae Jun and Sokbae Lee. Causal Inference in Case-Control Studies.
\url{https://arxiv.org/abs/2004.08318}.
}
