% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicc.R
\name{cicc}
\alias{cicc}
\title{Causal Inference for Relative Risk}
\usage{
cicc(y, t, x, p = 1L)
}
\arguments{
\item{y}{n-dimensional vector of binary outcomes}

\item{t}{n-dimensional vector of binary treatments}

\item{x}{n by p matrix of covariates}

\item{p}{a specified upper bound for the unknown true case probability (default = 1)}
}
\value{
An S3 object of type "ciccr". The object has the following elements.
\item{est}{a scalar estimate of the upper bound on the average of log relative risk}
\item{se}{standard error}
}
\description{
Provides an upper bound on the average of log relative risk
}
\examples{
# use the ACS dataset included in the package
  y = ciccr::ACS$topincome
  t = ciccr::ACS$baplus
  x = ciccr::ACS$age
# The upper bound for Pr(y=1) is set as p = 1 by default
  cicc(y, t, x)
# The upper bound for Pr(y=1) is set as p = 0.2 if a plausible upper bound is avaiable
  cicc(y, t, x, 0.2)
}
\references{
Sung Jae Jun and Sokbae Lee. Causal Inference in Case-Control Studies.
\url{https://arxiv.org/abs/2004.08318}.
}
