% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciccr.R
\docType{package}
\name{ciccr}
\alias{ciccr}
\title{ciccr: a package for causal inference in case-control studies}
\description{
The ciccr package provides methods for causal infernce in case-control studies.
}
\section{Functions}{

The package includes the following:
\itemize{
\item{\code{\link{cicc}}: }{carries out causual inference on relative risk.}
\item{\code{\link{avg_retro_logit}}: }{averages the log odds ratio using retrospective logistic regression.}
\item{\code{\link{cicc_AR}}: }{carries out causual inference on attributable risk.}
\item{\code{\link{avg_AR_logit}}: }{averages the upper bound of causal attributable risk using logistic regression models.}
\item{\code{\link{ACS}}: }{provides an illustrative case-control sample.}
}
}

\references{
Sung Jae Jun and Sokbae Lee. Causal Inference in Case-Control Studies.
\url{https://arxiv.org/abs/2004.08318}.
}
